updates to new base for feedback generator
