+++ notebook.py
@@ -4,7 +4,7 @@
 Automatically generated by Colab.
 
 Original file is located at
-    https://colab.research.google.com/drive/1uCATOAaJEKGIOoq3Tjn6qJelkQVIWfYh
+    https://colab.research.google.com/drive/1HrjNuHF4k-RXkz6nUfFMLTA61r9wp8eE
 
 # Main Feedback Generation Notebook
 This notebook handles preprocessing, model interaction, and feedback generation.
@@ -25,111 +25,19 @@ This notebook handles preprocessing, model interaction, and feedback generation.
 # %cd project
 
 # List all branches (optional, for checking)
-!git branch -a
+# !git branch -a
 
 # Checkout YOUR branch (!!replace "your-branch-name"!!)
-!git checkout -b scraping/inline_feedback origin/scraping/inline_feedback
-
-# Add your src folder to the Python path
-import sys
-sys.path.append('/content/project/')
-
-# import data path for loading files
-import os
-# Base data directory
-data_base = '/content/project/data'
-# Paths to specific subfolders
-raw_data_path = os.path.join(data_base, 'raw')
-processed_data_path = os.path.join(data_base, 'processed')
+!git checkout -b model/prompting origin/model/prompting
 
 """**--Install required libraries**"""
 
 # Commented out IPython magic to ensure Python compatibility.
-# %pip install -r requirements.txt
-!pip install python-docx
+# %pip install -r requirements.txt --quiet
 
+# Add src folder to python path to edit python files
 import sys
-import os
-import re
-import json
-from pathlib import Path
-from google.colab import drive
-
-# Mount Google Drive (uncomment if needed)
-# drive.mount('/content/drive')
-
-# Set up paths
-ROOT = '/content/project'
-sys.path.append(ROOT)
-
-# Import after adding to path
-from tropos.preprocess_docx.comments import Comments
-
-# Configure paths
-raw_dir = f"{ROOT}/data/raw/"
-processed_dir = f"{ROOT}/data/processed/student_submissions/"
-
-def process_student_submissions():
-    """Process all student submissions and save comments to structured directories."""
-    # Create processed directory if needed
-    Path(processed_dir).mkdir(parents=True, exist_ok=True)
-
-    all_comments = {}  # Master dictionary for backup
-
-    for student_folder in os.listdir(raw_dir):
-        student_path = os.path.join(raw_dir, student_folder)
-
-        if not os.path.isdir(student_path):
-            continue
-
-        print(f"\nProcessing {student_folder}...")
-
-        for filename in os.listdir(student_path):
-            if not filename.endswith('.docx'):
-                continue
-
-            # Parse student number and part number
-            student_num = re.search(r'Student[ _](\d+)', filename, re.IGNORECASE)
-            part_info = re.search(r'(Part[ _]?(\d+)|Final)', filename, re.IGNORECASE)
-
-            if not student_num or not part_info:
-                print(f"  Skipping {filename} - couldn't parse info")
-                continue
-
-            # Standardize naming
-            student_key = f"Student_{student_num.group(1).zfill(2)}"
-            part_key = part_info.group(0).lower().replace(' ', '')
-
-            # Set up target path
-            target_dir = os.path.join(processed_dir, student_key)
-            Path(target_dir).mkdir(exist_ok=True)
-
-            target_path = os.path.join(target_dir, f"{part_key}.json")
-            doc_path = os.path.join(student_path, filename)
-
-            # Extract and save comments
-            try:
-                comments = Comments(doc_path).parse_comments().get_results()
-
-                with open(target_path, 'w') as f:
-                    json.dump(comments, f, indent=2)
-
-                # Update master record
-                if student_key not in all_comments:
-                    all_comments[student_key] = {}
-                all_comments[student_key][part_key] = comments
-
-                print(f"  Processed: {filename} → {target_path}")
-
-            except Exception as e:
-                print(f"  Error processing {filename}: {str(e)}")
-
-
-
-    print(f"\nProcessing complete!")
-
-# Run the processing
-process_student_submissions()
+sys.path.append('/content/project/')
 
 """**--Import modules (youre working on)**
 <br>*each of our classes will be what will merge to this notebook (im p sure)*
@@ -138,5 +46,16 @@ process_student_submissions()
 <br> *the following is an example with my Rubric module*
 """
 
-# Import and call main functionfrom tropos import main
-main()
+from tropos import StudentSubmission
+from tropos.models import gpt
+
+example = StudentSubmission(
+    "project/data/raw/Student 1/Student 1 Part 1.docx",
+    "project/data/processed/assignment/assignment.docx")
+target = StudentSubmission(
+    "/project/data/unmarked_raw/Uncommented Student 1.docx",
+    "project/data/processed/assignment/assignment.docx")
+
+feedback = gpt.generate_feedback(example, target)
+print(feedback)
+
