+++ tropos/__init__.py
@@ -1,17 +1,14 @@
 # This file marks tropos as a Python package.
 
-from .models.gpt import generate_feedback
-from .models.trained import load_model
-
 from .preprocess_docx.rubric import parse_rubric
 from .preprocess_docx.submission import parse_submission
-from .preprocess_docx.requirements import parse_requirements
-from .preprocess_docx.comments import Comments, parse_comments
+from .preprocess_docx.assignment_requirements import parse_requirements
+from .preprocess_docx.comments import Comments
 from .preprocess_docx import StudentSubmission
 
 
 # Optional UI export
-from .gradio.ui import launch_ui
+# from .gradio.ui import launch_ui
 
 # Starts the program
 def main():
